Fix geometry import
